verify admin

+++ routes/dbs.js
@@ -6,36 +6,47 @@ const Shelf = require("../models/shelf");
 const User = require("../models/user");
 
 const passport = require("passport");
-const { initialize:initializePassport, isAdmin:isAdmin}  = require("../passportConfig");
+const {
+  initialize: initializePassport,
+  isAdmin: isAdmin,
+} = require("../passportConfig");
 initializePassport(passport);
 
 // Delete one
-router.delete("/:id",  passport.authenticate("jwt", { session: false }), async (req, res) => {
-  try {
-    switch (req.params.id) {
-      case "books":
-        await Book.collection.drop();
-        res.json({ message: "Deleted books collection" });
-        break;
-      case "shelves":
-        await Shelf.collection.drop();
-        res.json({ message: "Deleted shelves collection" });
-        break;
-      case "bookdetails":
-        await BookDetails.collection.drop();
-        res.json({ message: "Deleted bookDetailss collection" });
-        break;
-      case "users":
-        await User.collection.drop();
-        res.json({ message: "Deleted users collection" });
-        break;
-      default:
-        res.status(404).json({ message: "collection not found" });
+router.delete(
+  "/:id",
+  passport.authenticate("jwt", { session: false }),
+  async (req, res) => {
+    if (!isAdmin(req)) {
+      res.status(403).json({ message: "insufficient righs" });
+      return;
     }
-  } catch (err) {
-    res.status(500).json({ message: err.message });
+    try {
+      switch (req.params.id) {
+        case "books":
+          await Book.collection.drop();
+          res.json({ message: "Deleted books collection" });
+          break;
+        case "shelves":
+          await Shelf.collection.drop();
+          res.json({ message: "Deleted shelves collection" });
+          break;
+        case "bookdetails":
+          await BookDetails.collection.drop();
+          res.json({ message: "Deleted bookDetailss collection" });
+          break;
+        case "users":
+          // await User.collection.drop();
+          res.json({ message: "Deleted users collection" });
+          break;
+        default:
+          res.status(404).json({ message: "collection not found" });
+      }
+    } catch (err) {
+      res.status(500).json({ message: err.message });
+    }
   }
-});
+);
 
 module.exports = router;
